created workspace for showing algorithms

+++ lib/sequence_gen/random_walk.cpp
@@ -1,6 +1,6 @@
 #include "random_walk.h"
 
-RandomWalk::RandomWalk( std::function<float()> next_incr_generator) : m_gen_next_incr(next_incr_generator)
+RandomWalk::RandomWalk( std::function<double()> next_incr_generator) : m_gen_next_incr(next_incr_generator)
 {
   m_walk.push_back(0.0);
 }
@@ -12,6 +12,6 @@ void RandomWalk::gen_steps(unsigned int n)
   }
 }
 
-const deque<float>& RandomWalk::get_walk() {
+const deque<double>& RandomWalk::get_walk() {
   return m_walk;
 }
